phase1_knowledgeBase/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# Vector DB & Embeddings
pinecone>=5.0.1

# LLM
openai==1.10.0
httpx==0.26.0

# Utilities
python-json-logger==2.0.7
structlog==24.1.0
tenacity==8.2.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
